docs(FR-08): AI レビュー対応（platform ADR 参照の曖昧さ解消・逐次保存の意図明記・float→double 明示）
- microservices-platform 側 ADR（IADR-0044/0047）を本リポ IADR-0044/0047 と別採番と明記し曖昧さを解消（🟡）。
  IADR-0069 に跨ぎ注意の脚注を追加（実在の確認元ファイルも明記）。
- KnowledgeBaseWriterSink の逐次保存が意図的選択である旨と #9 でのバッチ/並列検討をコメント化（🟡）。
- HttpKnowledgeBaseSearch の score を float→double 明示変換（拡大・非損失）にしコメント付与（🟢）。

Refs #18

+++ backend/Shared/AiStockTrading.Shared.KnowledgeBase/Adapters/HttpKnowledgeBaseSearch.cs
@@ -66,7 +66,7 @@ internal sealed class HttpKnowledgeBaseSearch(
                     r.DocumentId,
                     r.DocumentTitle,
                     r.Text,
-                    r.Score,
+                    (double)r.Score, // platform は float スコア。KnowledgeHit は double のため明示変換（拡大・非損失）。
                     r.MarkdownUri,
                     r.Tags ?? []))
                 .ToList();
